day01 - implementation done

+++ day01/src/main/java/org/tekeli/borisp/adventcode2023/day01/CalibrationSum.java
@@ -1,4 +1,4 @@
 package org.tekeli.borisp.adventcode2023.day01;
 
-public record CalibrationSum(Long value) {
+public record CalibrationSum(Integer value) {
 }
